refactor(tests): consolidate duplicate helpers into conftest.py
Moves bootstrap_database, insert_event_row, insert_job_row,
build_signature, and clear_logger into the canonical conftest.py.
Removes all per-file duplicates (9 files updated). Adds a comment
directing contributors to check conftest before adding new helpers.

Also catches test_app.py's bootstrap_database copy missed in the plan.
