+++ tests/conftest.py
@@ -1,6 +1,14 @@
 import asyncio
+import base64
+import hashlib
+import hmac
+import logging
+from pathlib import Path
 
+from dmguard.job_machine import JobStage, JobStatus
 from dmguard.secrets import SecretStore
+
+# Shared test helpers — check here before writing a new one in a test file.
 
 
 class StubSecretStore(SecretStore):
@@ -18,3 +26,80 @@ class StubSecretStore(SecretStore):
 
 def run(coroutine):
     return asyncio.run(coroutine)
+
+
+async def bootstrap_database(db_path: Path) -> None:
+    from dmguard.db import get_connection
+    from dmguard.schema import bootstrap_schema
+
+    async with get_connection(db_path) as connection:
+        await bootstrap_schema(connection)
+
+
+async def insert_event_row(
+    db_path: Path,
+    *,
+    event_id: str,
+    sender_id: str | None = None,
+) -> None:
+    from dmguard.db import get_connection
+    from dmguard.repo_events import insert_event
+
+    async with get_connection(db_path) as connection:
+        await insert_event(
+            connection,
+            event_id=event_id,
+            received_at="2026-03-11T00:00:00Z",
+            payload_json='{"event_id":"%s"}' % event_id,
+            sender_id=sender_id,
+        )
+        await connection.commit()
+
+
+async def insert_job_row(
+    db_path: Path,
+    *,
+    event_id: str,
+    next_run_at: str,
+    status: JobStatus = JobStatus.queued,
+    stage: JobStage = JobStage.fetch_dm,
+    attempt: int = 0,
+    processing_started_at: str | None = None,
+) -> int:
+    from dmguard.db import get_connection
+    from dmguard.repo_jobs import insert_job
+
+    async with get_connection(db_path) as connection:
+        job_id = await insert_job(
+            connection,
+            event_id=event_id,
+            status=status,
+            stage=stage,
+            next_run_at=next_run_at,
+            attempt=attempt,
+            processing_started_at=processing_started_at,
+        )
+        await connection.commit()
+
+    return job_id
+
+
+def build_signature(raw_body: bytes, consumer_secret: str) -> str:
+    digest = hmac.new(
+        consumer_secret.encode("utf-8"),
+        raw_body,
+        hashlib.sha256,
+    ).digest()
+    encoded = base64.b64encode(digest).decode("ascii")
+    return f"sha256={encoded}"
+
+
+def clear_logger(name: str) -> None:
+    logger = logging.getLogger(name)
+
+    for handler in list(logger.handlers):
+        logger.removeHandler(handler)
+        handler.close()
+
+    logger.setLevel(logging.NOTSET)
+    logger.propagate = True
