+++ tests/test_repo.py
@@ -4,15 +4,7 @@ from pathlib import Path
 import pytest
 
 from dmguard.job_machine import JobStage, JobStatus
-from tests.conftest import run
-
-
-async def bootstrap_database(db_path: Path) -> None:
-    from dmguard.db import get_connection
-    from dmguard.schema import bootstrap_schema
-
-    async with get_connection(db_path) as connection:
-        await bootstrap_schema(connection)
+from tests.conftest import bootstrap_database, insert_event_row, insert_job_row, run
 
 
 async def fetch_row(
@@ -25,54 +17,6 @@ async def fetch_row(
         row = await cursor.fetchone()
 
     return row
-
-
-async def insert_event_row(
-    db_path: Path,
-    *,
-    event_id: str,
-    sender_id: str | None = None,
-) -> None:
-    from dmguard.db import get_connection
-    from dmguard.repo_events import insert_event
-
-    async with get_connection(db_path) as connection:
-        await insert_event(
-            connection,
-            event_id=event_id,
-            received_at="2026-03-11T00:00:00Z",
-            payload_json='{"event_id":"%s"}' % event_id,
-            sender_id=sender_id,
-        )
-        await connection.commit()
-
-
-async def insert_job_row(
-    db_path: Path,
-    *,
-    event_id: str,
-    next_run_at: str,
-    status: JobStatus = JobStatus.queued,
-    stage: JobStage = JobStage.fetch_dm,
-    attempt: int = 0,
-    processing_started_at: str | None = None,
-) -> int:
-    from dmguard.db import get_connection
-    from dmguard.repo_jobs import insert_job
-
-    async with get_connection(db_path) as connection:
-        job_id = await insert_job(
-            connection,
-            event_id=event_id,
-            status=status,
-            stage=stage,
-            next_run_at=next_run_at,
-            attempt=attempt,
-            processing_started_at=processing_started_at,
-        )
-        await connection.commit()
-
-    return job_id
 
 
 def test_repo_events_insert_and_get(tmp_path: Path) -> None:
